% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposureApp.R
\name{exposureApp}
\alias{exposureApp}
\title{Launch the Species Climate Exposure Shiny App}
\usage{
exposureApp(
  intDir = NULL,
  launch.browser = TRUE,
  host = "127.0.0.1",
  port = getOption("shiny.port")
)
}
\arguments{
\item{intDir}{Path to the directory containing the app data files
(\code{allExpForShiny.qs}, \code{landTemplate.tif}, etc.).
Defaults to the package's built-in \code{extdata} directory.}

\item{launch.browser}{Logical; open the app in the default browser. Default \code{TRUE}.}

\item{host}{Host IP address to listen on. Default \code{"127.0.0.1"}.}

\item{port}{Port number. Default uses \code{getOption("shiny.port")}.}
}
\value{
Launches a Shiny application (does not return a value).
}
\description{
Launch the Species Climate Exposure Shiny App
}
